requirements.txt
numpy
opencv-python-headless
onnxruntime-gpu
pillow
streamlit
torchvision